refactor(automations): vertical-slice actions and triggers by domain
