--- surfsense_backend/app/automations/schemas/actions/__init__.py
@@ -1,9 +1,0 @@
-"""Per-action params schemas, one per action type."""
-
-from __future__ import annotations
-
-from .agent_task import AgentTaskActionParams
-
-__all__ = [
-    "AgentTaskActionParams",
-]
