--- surfsense_backend/app/automations/schemas/triggers/__init__.py
@@ -1,11 +1,0 @@
-"""Per-trigger params schemas, one per trigger type."""
-
-from __future__ import annotations
-
-from .manual import ManualTriggerParams
-from .schedule import ScheduleTriggerParams
-
-__all__ = [
-    "ManualTriggerParams",
-    "ScheduleTriggerParams",
-]
