+++ surfsense_backend/app/automations/triggers/types.py
